ch. 02: tested with PostgreSQL.

+++ trunk/sbia/ch02/src/main/resources/create-tables.sql
@@ -4,6 +4,6 @@ CREATE TABLE invoice
   customer_id integer NOT NULL,
   description character varying(50),
   issue_date date,
-  amount double,
+  amount float,
   CONSTRAINT invoice_pkey PRIMARY KEY (id)
 );
